Warnings suppression
git-svn-id: svn://svn.code.sf.net/p/mediainfo/code/MediaInfoLib/trunk@313 45c5ae6f-87cc-4fd0-92ee-e4f023fd80da

+++ Source/MediaInfo/Video/File_Vc1.cpp
@@ -376,7 +376,7 @@ void File_Vc1::FrameHeader_Fill()
     if (framerate_present)
     {
         if (framerate_form)
-            FrameRate=((float32)(framerateexp+1))/32.0;
+            FrameRate=((float32)(framerateexp+1))/(float32)32;
         else if (Vc1_FrameRate_dr(frameratecode_dr))
             FrameRate=Vc1_FrameRate_enr(frameratecode_enr)/Vc1_FrameRate_dr(frameratecode_dr);
     }
